Ensure cluster click causes an increase in zoom
When setting Google Maps to show a boundary it tries to maintain spacing
between that boundary and the edge of the container. This can be a pain
on small screens because we're short on screen space already. we need to
store the maps current zoom level and then test it once the boundary has
been set. If the zoom level has not been changed then increase the zoom
level manually